+++ packages/miniflare/src/index.ts
@@ -2,6 +2,7 @@ import assert from "assert";
 import crypto from "crypto";
 import { Abortable } from "events";
 import fs from "fs";
+import { mkdir, writeFile } from "fs/promises";
 import http from "http";
 import net from "net";
 import os from "os";
@@ -56,6 +57,7 @@ import {
 	QueuesError,
 	R2_PLUGIN_NAME,
 	ReplaceWorkersTypes,
+	SECRET_STORE_PLUGIN_NAME,
 	SERVICE_ENTRY,
 	SharedOptions,
 	SOCKET_ENTRY,
@@ -109,6 +111,7 @@ import {
 	SharedHeaders,
 	SiteBindings,
 } from "./workers";
+import { ADMIN_API } from "./workers/secrets-store/constants";
 import { formatZodError } from "./zod-format";
 import type {
 	CacheStorage,
@@ -116,6 +119,7 @@ import type {
 	DurableObjectNamespace,
 	Fetcher,
 	KVNamespace,
+	KVNamespaceListKey,
 	Queue,
 	R2Bucket,
 } from "@cloudflare/workers-types/experimental";
@@ -979,6 +983,17 @@ export class Miniflare {
 				if (!colors$.enabled) message = stripAnsi(message);
 				this.#log.logWithLevel(logLevel, message);
 				response = new Response(null, { status: 204 });
+			} else if (url.pathname === "/core/store-temp-file") {
+				const prefix = url.searchParams.get("prefix");
+				const folder = prefix ? `files/${prefix}` : "files";
+				await mkdir(path.join(this.#tmpPath, folder), { recursive: true });
+				const filePath = path.join(
+					this.#tmpPath,
+					folder,
+					`${crypto.randomUUID()}.${url.searchParams.get("extension") ?? "txt"}`
+				);
+				await writeFile(filePath, await request.text());
+				response = new Response(filePath, { status: 200 });
 			}
 		} catch (e: any) {
 			this.#log.error(e);
@@ -1544,11 +1559,13 @@ export class Miniflare {
 			const ready = initial ? "Ready" : "Updated and ready";
 
 			const urlSafeHost = getURLSafeHost(configuredHost);
-			this.#log.info(
-				`${ready} on ${secure ? "https" : "http"}://${urlSafeHost}:${entryPort}`
-			);
+			if (this.#sharedOpts.core.logRequests) {
+				this.#log.info(
+					`${ready} on ${secure ? "https" : "http"}://${urlSafeHost}:${entryPort}`
+				);
+			}
 
-			if (initial) {
+			if (initial && this.#sharedOpts.core.logRequests) {
 				const hosts: string[] = [];
 				if (configuredHost === "::" || configuredHost === "*") {
 					hosts.push("localhost");
@@ -1911,6 +1928,34 @@ export class Miniflare {
 		workerName?: string
 	): Promise<ReplaceWorkersTypes<KVNamespace>> {
 		return this.#getProxy(KV_PLUGIN_NAME, bindingName, workerName);
+	}
+	getSecretsStoreSecretAPI(
+		bindingName: string,
+		workerName?: string
+	): Promise<
+		() => {
+			create: (value: string) => Promise<string>;
+			update: (value: string, id: string) => Promise<string>;
+			duplicate: (id: string, newName: string) => Promise<string>;
+			delete: (id: string) => Promise<void>;
+			list: () => Promise<KVNamespaceListKey<{ uuid: string }, string>[]>;
+			get: (id: string) => Promise<string>;
+		}
+	> {
+		return this.#getProxy(
+			SECRET_STORE_PLUGIN_NAME,
+			bindingName,
+			workerName
+		).then((binding) => {
+			// @ts-expect-error We exposed an admin API on this key
+			return binding[ADMIN_API];
+		});
+	}
+	getSecretsStoreSecret(
+		bindingName: string,
+		workerName?: string
+	): Promise<ReplaceWorkersTypes<KVNamespace>> {
+		return this.#getProxy(SECRET_STORE_PLUGIN_NAME, bindingName, workerName);
 	}
 	getQueueProducer<Body = unknown>(
 		bindingName: string,