+++ packages/miniflare/src/plugins/core/index.ts
@@ -210,6 +210,10 @@ export const CoreSharedOptionsSchema = z.object({
 	unsafeModuleFallbackService: ServiceFetchSchema.optional(),
 	// Keep blobs when deleting/overwriting keys, required for stacked storage
 	unsafeStickyBlobs: z.boolean().optional(),
+	// Enable directly triggering user Worker handlers with paths like `/cdn-cgi/handler/scheduled`
+	unsafeTriggerHandlers: z.boolean().optional(),
+	// Enable logging requests
+	logRequests: z.boolean().default(true),
 });
 
 export const CORE_PLUGIN_NAME = "core";
@@ -746,6 +750,14 @@ export function getGlobalServices({
 	const serviceEntryBindings: Worker_Binding[] = [
 		WORKER_BINDING_SERVICE_LOOPBACK, // For converting stack-traces to pretty-error pages
 		{ name: CoreBindings.JSON_ROUTES, json: JSON.stringify(routes) },
+		{
+			name: CoreBindings.TRIGGER_HANDLERS,
+			json: JSON.stringify(!!sharedOptions.unsafeTriggerHandlers),
+		},
+		{
+			name: CoreBindings.LOG_REQUESTS,
+			json: JSON.stringify(!!sharedOptions.logRequests),
+		},
 		{ name: CoreBindings.JSON_CF_BLOB, json: JSON.stringify(sharedOptions.cf) },
 		{ name: CoreBindings.JSON_LOG_LEVEL, json: JSON.stringify(log.level) },
 		{
@@ -795,13 +807,8 @@ export function getGlobalServices({
 			name: SERVICE_ENTRY,
 			worker: {
 				modules: [{ name: "entry.worker.js", esModule: SCRIPT_ENTRY() }],
-				compatibilityDate: "2023-04-04",
-				compatibilityFlags: [
-					"nodejs_compat",
-					"service_binding_extra_handlers",
-					"brotli_content_encoding",
-					"rpc",
-				],
+				compatibilityDate: "2025-03-17",
+				compatibilityFlags: ["nodejs_compat", "service_binding_extra_handlers"],
 				bindings: serviceEntryBindings,
 				durableObjectNamespaces: [
 					{