+++ packages/miniflare/src/workers/secrets-store/constants.ts
@@ -1,0 +1,1 @@
+export const ADMIN_API = "SecretsStoreSecret::admin_api";