+++ packages/wrangler/src/api/startDevWorker/types.ts
@@ -21,6 +21,7 @@ import type {
 	CfQueue,
 	CfR2Bucket,
 	CfScriptFormat,
+	CfSecretsStoreSecrets,
 	CfSendEmailBindings,
 	CfService,
 	CfUnsafe,
@@ -260,6 +261,7 @@ export type Binding =
 	| ({ type: "dispatch_namespace" } & BindingOmit<CfDispatchNamespace>)
 	| ({ type: "mtls_certificate" } & BindingOmit<CfMTlsCertificate>)
 	| ({ type: "pipeline" } & BindingOmit<CfPipeline>)
+	| ({ type: "secrets_store_secret" } & BindingOmit<CfSecretsStoreSecrets>)
 	| ({ type: "logfwdr" } & NameOmit<CfLogfwdrBinding>)
 	| { type: `unsafe_${string}` }
 	| { type: "assets" };