+++ packages/wrangler/src/deployment-bundle/worker.ts
@@ -87,11 +87,17 @@ export interface CfKvNamespace {
 /**
  * A binding to send email.
  */
-export interface CfSendEmailBindings {
+export type CfSendEmailBindings = {
 	name: string;
-	destination_address?: string;
-	allowed_destination_addresses?: string[];
-}
+} & (
+	| { destination_address?: string }
+	| { allowed_destination_addresses?: string[] }
+);
+// export interface CfSendEmailBindings {
+// 	name: string;
+// 	destination_address?: string | undefined;
+// 	allowed_destination_addresses?: string[] | undefined;
+// }
 
 /**
  * A binding to a wasm module (in service-worker format)
@@ -192,6 +198,12 @@ export interface CfD1Database {
 export interface CfVectorize {
 	binding: string;
 	index_name: string;
+}
+
+export interface CfSecretsStoreSecrets {
+	binding: string;
+	store_id: string;
+	secret_name: string;
 }
 
 export interface CfHyperdrive {
@@ -339,6 +351,7 @@ export interface CfWorkerInit {
 		d1_databases: CfD1Database[] | undefined;
 		vectorize: CfVectorize[] | undefined;
 		hyperdrive: CfHyperdrive[] | undefined;
+		secrets_store_secrets: CfSecretsStoreSecrets[] | undefined;
 		services: CfService[] | undefined;
 		analytics_engine_datasets: CfAnalyticsEngineDataset[] | undefined;
 		dispatch_namespaces: CfDispatchNamespace[] | undefined;